Fix for bela warning

+++ source/modules/soul_patch_loader/classes/soul_patch_BelaTransformation.h
@@ -70,7 +70,7 @@ namespace soul::patch
                     connections << name << "Input.audioOut -> wrappedModule." << name << ";" << newLine;
                     connections << newLine;
 
-                    nextOutputChannel += channels;
+                    nextOutputChannel += static_cast<uint32_t> (channels);
                 }
                 else
                 {
